[FIX] account_payment: ensure tab switching works even when no providers are available
