+++ addons/account_payment/static/src/js/payment_form.js
@@ -1,15 +1,8 @@
+/** @odoo-module **/
+
 import PaymentForm from "@payment/js/payment_form";
 
 PaymentForm.include({
-
-    /**
-     * @override
-     */
-    async start() {
-        await this._super(...arguments);
-        this._onChangePaymentTabs();
-    },
-
     /**
      * Set whether we are paying an installment before submitting.
      *
@@ -48,34 +41,4 @@ PaymentForm.include({
             transactionRouteParams.payment_reference = this.paymentContext.paymentReference;
             return transactionRouteParams;
         },
-
-    /**
-     * Handles payment tab changes (installment or full amount).
-     *
-     * This method listens for the `shown.bs.tab` event on payment tab buttons.
-     * When the user switches tabs, it updates the URL parameters `mode` and
-     * `render_change`, then reloads the page. This forces the backend to
-     * re-render the payment form with updated data, including the corresponding
-     * amount and available payment providers.
-     *
-     * Added URL parameters:
-     * - mode: either 'installment' or 'full', depending on the selected tab.
-     * - render_change: 'true', indicating that the change should trigger a re-render.
-     */
-    _onChangePaymentTabs() {
-        $('.o_btn_payment_tab').on('shown.bs.tab', function (event) {
-            const activatedTab = event.target.id;
-            let mode = (activatedTab === 'o_payment_installments_tab')
-                ? 'installment'
-                : (activatedTab === 'o_payment_full_tab' ? 'full' : false);
-
-            if (mode) {
-                const url = new URL(window.location.href);
-                url.searchParams.set('mode', mode);
-                url.searchParams.set('render_change', 'true');
-
-                window.location.href = url.toString();
-            }
-        });
-    }
 });
