+++ addons/account_payment/static/src/js/payment_installments.js
@@ -1,0 +1,47 @@
+import publicWidget from '@web/legacy/js/public/public_widget';
+
+publicWidget.registry.PaymentInstallments = publicWidget.Widget.extend({
+    selector: '#o_payment_installments_modal',
+
+    /**
+     * @override
+     */
+    async start() {
+
+        await this._super(...arguments);
+        this._onChangePaymentTabs();
+
+    },
+
+    /**
+     * Handles payment tab changes (installment or full amount).
+     *
+     * This method listens for the `shown.bs.tab` event on payment tab buttons.
+     * When the user switches tabs, it updates the URL parameters `mode` and
+     * `render_change`, then reloads the page. This forces the backend to
+     * re-render the payment form with updated data, including the corresponding
+     * amount and available payment providers.
+     *
+     * Added URL parameters:
+     * - mode: either 'installment' or 'full', depending on the selected tab.
+     * - render_change: 'true', indicating that the change should trigger a re-render.
+     */
+    _onChangePaymentTabs() {
+        $('.o_btn_payment_tab').on('shown.bs.tab', function (event) {
+            const activatedTab = event.target.id;
+            let mode = (activatedTab === 'o_payment_installments_tab')
+                ? 'installment'
+                : (activatedTab === 'o_payment_full_tab' ? 'full' : false);
+
+            if (mode) {
+                const url = new URL(window.location.href);
+                url.searchParams.set('mode', mode);
+                url.searchParams.set('render_change', 'true');
+
+                window.location.href = url.toString();
+            }
+        });
+    }
+})
+
+export default publicWidget.registry.PaymentInstallments;
